requirements.txt
ag2[openai]>=0.2
ddgs>=9.0
tavily-python>=0.5
pypdf>=4.0
python-docx>=1.0
python-pptx>=0.6.21
httpx>=0.27
beautifulsoup4>=4.12
fastapi>=0.109
python-multipart>=0.0.9
uvicorn[standard]>=0.27
pydantic>=2.0
youtube-transcript-api>=1.2.0
